fix: remove extra file and add functions to controllers
