+++ VAS/Areas/VAS/Controllers/TotalAmountReceivedThisMonthController.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
+using System.Globalization;
 using System.Web.Mvc;
 using VAdvantage.Classes;
 using VAdvantage.DataBase;
@@ -70,8 +71,8 @@ namespace VIS.Controllers
                 WHERE Payment.IsReceipt = 'Y'
                   AND Payment.IsActive = 'Y'
                   AND Payment.DocStatus IN ('CO', 'CL')
-                  AND " + WidgetDateSqlHelper.TruncColumn("Payment.DateAcct") + @" >= " + WidgetDateSqlHelper.ToSqlDate(monthStart) + @"
-                  AND " + WidgetDateSqlHelper.TruncColumn("Payment.DateAcct") + @" < " + WidgetDateSqlHelper.ToSqlDate(nextMonthStart);
+                  AND " + TruncColumn("Payment.DateAcct") + @" >= " + ToSqlDate(monthStart) + @"
+                  AND " + TruncColumn("Payment.DateAcct") + @" < " + ToSqlDate(nextMonthStart);
 
             receivedThisMonthSql = MRole.GetDefault(ctx).AddAccessSQL(
                 receivedThisMonthSql,
@@ -134,6 +135,32 @@ namespace VIS.Controllers
                     dr.Dispose();
                 }
             }
+        }
+
+
+
+        internal static string ToSqlDate(DateTime date)
+        {
+            DateTime day = date.Date;
+
+            if (DB.IsOracle())
+            {
+                return "TO_DATE('"
+                    + day.ToString("yyyy-MM-dd", CultureInfo.InvariantCulture)
+                    + "','YYYY-MM-DD')";
+            }
+
+            return DB.TO_DATE(day, true);
+        }
+
+        internal static string TruncColumn(string columnExpression)
+        {
+            if (DB.IsOracle())
+            {
+                return "TRUNC(" + columnExpression + ")";
+            }
+
+            return columnExpression;
         }
     }
 }
